requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
oauthlib==3.2.2
packaging==23.1
pluggy==1.0.0
praw==7.7.0
prawcore==2.3.0
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.3.1
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.15
websocket-client==1.5.1
